requirements.txt
torch
gradio
matplotlibmatplotlib
